moved primordial wave to talents
