+++ src/analysis/retail/shaman/restoration/CombatLogParser.tsx
@@ -1,11 +1,4 @@
-import {
-  AstralShift,
-  ChainHarvest,
-  ElementalConduit,
-  SpiritWolf,
-  StaticCharge,
-  TumblingWaves,
-} from 'analysis/retail/shaman/shared';
+import { AstralShift, SpiritWolf, StaticCharge } from 'analysis/retail/shaman/shared';
 import CoreCombatLogParser from 'parser/core/CombatLogParser';
 import ManaTracker from 'parser/core/healingEfficiency/ManaTracker';
 import LowHealthHealing from 'parser/shared/modules/features/LowHealthHealing';
@@ -26,14 +19,7 @@ import MasteryEffectiveness from './modules/features/MasteryEffectiveness';
 import SpellUsable from './modules/features/SpellUsable';
 import TidalWaves from './modules/features/TidalWaves';
 // Talents
-import EmbraceOfEarth from './modules/shadowlands/conduits/EmbraceOfEarth';
-import HeavyRainfall from './modules/shadowlands/conduits/HeavyRainfall';
-import NaturesFocus from './modules/shadowlands/conduits/NaturesFocus';
-import SwirlingCurrents from './modules/shadowlands/conduits/SwirlingCurrents';
-import EarthenHarmony from './modules/shadowlands/legendaries/EarthenHarmony';
-import JonatsNaturalFocus from './modules/shadowlands/legendaries/JonatsNaturalFocus';
-import PrimalTideCore from './modules/shadowlands/legendaries/PrimalTideCore';
-import PrimordialWave from './modules/shadowlands/spells/PrimordialWave';
+import PrimordialWave from './modules/talents/PrimordialWave';
 import ChainHeal from './modules/spells/ChainHeal';
 import EarthShield from './modules/spells/EarthShield'; // technically shared
 import HealingRain from './modules/spells/HealingRain';
@@ -107,6 +93,7 @@ class CombatLogParser extends CoreCombatLogParser {
     naturesGuardian: NaturesGuardian,
     ancestralProtectionTotem: AncestralProtectionTotem,
     talentStatisticBox: TalentStatisticBox,
+    primordialWave: PrimordialWave,
 
     // Spells
     chainHeal: ChainHeal,
@@ -128,23 +115,6 @@ class CombatLogParser extends CoreCombatLogParser {
     // Normalizers
     cloudburstNormalizer: CloudburstNormalizer,
     riptideNormalizer: RiptideNormalizer,
-
-    // Conduits
-    embraceOfEarth: EmbraceOfEarth,
-    heavyRainfall: HeavyRainfall,
-    swirlingCurrents: SwirlingCurrents,
-    naturesFocus: NaturesFocus,
-    tumblingWaves: TumblingWaves,
-
-    // Legendaries
-    primalTideCore: PrimalTideCore,
-    jonatsNaturalFocus: JonatsNaturalFocus,
-    earthenHarmony: EarthenHarmony,
-    elementalConduit: ElementalConduit,
-
-    // Covenants
-    chainHarvest: ChainHarvest,
-    primordialWave: PrimordialWave,
   };
 }
